Remove redundant comment in Container\Local::ensureDirectory()
The comment in ensureDirectory() refers to the fact that in the past the
code used to set the mode to 766, but currently there is no sign of that
in the code.

The info about the history of the change is already in the commit
message of commit e44ec1b482f6 (fix prefix directory permissions), so
there is no need to repeat that in the code.