Create the Container directory only if there are no collisions
Checking existence with is_dir() does not cover the case when the file
exists but is not a directory.

So use file_exists() and create the directory only if there is no
collision with an existing file with the same name, regardless of the
file type.

The code checks below if the file is actually a directory to communicate
this info to the caller, so it's safe no to do that before creating the
directory.