Make Util/Arr::normalizeFiles() more robust
The caller may pass either $_FILES['some_field_name'] or $_FILES, and in
either case there could be a single file or multiple files.

So cover all the four cases in a more rigorous way, also handling the
case when the $files array does not need to be remapped.

NOTE, use $files[$keys[0]] instead of $files[0] because when passing
$_FILES the array is associative and the numerical key does not work.